# Add C-string overloads to String class