… tests/core/string/test_string.h