@Rubonnek 2023-11-03
I noticed I missed another overload I wanted and took the liberty to add `int String::findn(const char *p_str, int p_from)` as well, which is the last one I'll add for this PR.