@AThousandShips 2023-11-03
Let's add some unit tests for these to ensure correctness, do they work with unicode? If so then ensure it works with non-utf8 input