@AThousandShips 2023-11-03
It currently does break current checks, and unsure what happens with various cases using non-ascii/utf8 input