@Rubonnek 2023-11-03
> Edit: or rather, you need to rework the tests to test for the String and c-string versions, because the current tests use string literals that are now going through a different method

Indeed. I was about to ask about this -- I'm not sure if we want to duplicate the tests like this:

```cpp
	CHECK(s.findn("WHA") == 7);
	CHECK(s.findn(String("WHA")) == 7);
```

Or maybe add a macro to test both under the same line and keep it DRY.

I don't know if there's a cleaner way to do this. There doesn't seem to be.