@Rubonnek 2023-11-03
> would be good to confirm that it doesn't break things like: str.findn(U"...")

I should clarify I intended to keep this PR unrelated to `U""` and `L""` C-strings. Those calls won't break due to this PR because they are on an unrelated code path. The call `str.findn(U"")` yields a function signature similar to `findn(const char32_t*)`  for which there is no implementation (this PR only implements `findn(const char *)`) and as a result the compiler will look for other possible function candidates it can call given the `findn(const char32_t*)` signature. Currently, there's only one other candidate which is `str.findn(String)` which the compiler knows how to call because it can convert the `const char32_t*` parameter to a `String` through the `String` copy constructor `String::String(const char32_t*)`. The same happens with a `str.findn(L"")` function call and, again, this PR never touches either of those code paths.

Having said that, we are indeed missing some tests for `U""` and `L""` C-strings. I'll add the tests for the functions I've overloaded if they pass.