@AThousandShips 2023-11-03
Agreed, just wanted to make sure the automatic overload resolution works correctly as C++ can be so messy with that some times