@Rubonnek 2023-11-03
Tests passed for `U""` and `L""` C-strings on my end so I've added them.