@akien-mga 2024-01-04
> I personally find that this patch makes the Editor feel snappier, but I have not run any benchmarks yet.

Would be nice to do some benchmarking on an optimized build to see if there is actually a significant difference. Performance optimization can have surprising results only reasoned about but not tested.

CC @Calinou