@Calinou 2024-01-06
> Would be nice to do some benchmarking on an optimized build to see if there is actually a significant difference. Performance optimization can have surprising results only reasoned about but not tested.

Here benchmarks of editor startup + shutdown time (I ran these twice and results were consistent):

```
❯ hyperfine -iw1 "bin/godot.linuxbsd.editor.x86_64.master --path /tmp/4 -e --quit" "bin/godot.linuxbsd.editor.x86_64 --path /tmp/4 -e --quit"
Benchmark 1: bin/godot.linuxbsd.editor.x86_64.master --path /tmp/4 -e --quit
  Time (mean ± σ):      3.027 s ±  0.058 s    [User: 2.075 s, System: 0.361 s]
  Range (min … max):    2.911 s …  3.132 s    10 runs
 
Benchmark 2: bin/godot.linuxbsd.editor.x86_64 --path /tmp/4 -e --quit           <--- This PR
  Time (mean ± σ):      2.982 s ±  0.290 s    [User: 2.074 s, System: 0.366 s]
  Range (min … max):    2.523 s …  3.537 s    10 runs
 
Summary
  bin/godot.linuxbsd.editor.x86_64 --path /tmp/4 -e --quit ran
    1.02 ± 0.10 times faster than bin/godot.linuxbsd.editor.x86_64.master --path /tmp/4 -e --quit
```

Binary size:

### Editor

```
master: 105,335,696 bytes
    PR: 105,270,128 bytes (-65 KB)
```

### Release export template

```
master: 62,385,544 bytes
    PR: 62,344,616 bytes (-41 KB)
```