@Rubonnek 2024-03-11
Rebased to incorporate test changes from #89194.